Revert "Use molblock native wedging where available" (#1753)
Revert "Use molblock native wedging where available (#1731)"

This reverts commit 79c17ca08045a6da649ec3c76c4419979d9abaaf.

+++ packages/Chem/src/utils/mol-creation_rdkit.ts
@@ -4,7 +4,6 @@ export interface IMolContext {
   mol: RDMol | null; // null when molString is invalid
   kekulize: boolean;
   isQMol: boolean;
-  useMolBlockWedging: boolean;
 }
 
 export function isSmarts(molString: string): boolean {
@@ -14,7 +13,6 @@ export function isSmarts(molString: string): boolean {
 export function getMolSafe(molString: string, details: object = {}, rdKitModule: RDModule, warnOff: boolean = true): IMolContext {
   let isQMol = false;
   let kekulize: boolean = true;
-  let useMolBlockWedging: boolean = false;
   let mol: RDMol | null = null;
 
   try {
@@ -22,7 +20,7 @@ export function getMolSafe(molString: string, details: object = {}, rdKitModule:
     mol = _isSmarts ? rdKitModule.get_qmol(molString) : rdKitModule.get_mol(molString, JSON.stringify(details)); 
   }
   catch (e) {
-    if (mol !== null) {
+    if (mol !== null && mol.is_valid()) {
       mol.delete();
       mol = null;
     }
@@ -30,30 +28,26 @@ export function getMolSafe(molString: string, details: object = {}, rdKitModule:
     kekulize = false;
     try { mol = rdKitModule.get_mol(molString, JSON.stringify({ ...details, kekulize })); }
     catch (e2) {
-      if (mol !== null) {
+      if (mol !== null && mol.is_valid()) {
         mol.delete();
         mol = null;
       }
 
       try { mol = rdKitModule.get_qmol(molString); }
       catch (e3) {
-        if (mol !== null) {
+        if (mol !== null && mol.is_valid()) {
           mol.delete();
           mol = null;
         }
+        mol = null;
         if (!warnOff)
           console.error('Chem | In getMolSafe: RDKit.get_mol crashes on a molString: `' + molString + '`');
-        return { mol, kekulize, isQMol, useMolBlockWedging };
+        return { mol, kekulize, isQMol };
       }
-      return { mol, kekulize, isQMol, useMolBlockWedging };
+      isQMol = true;
+      return { mol, kekulize, isQMol };
     }
-    if (mol.is_valid()) {
-      useMolBlockWedging = mol.has_coords();
-    }
-    return { mol, kekulize, isQMol, useMolBlockWedging };
+    return { mol, kekulize, isQMol };
   }
-  if (mol.is_valid()) {
-    useMolBlockWedging = mol.has_coords();
-  }
-  return { mol, kekulize, isQMol, useMolBlockWedging };
+  return { mol, kekulize, isQMol };
 }
